Add forbidben redirects for non ajax requests

+++ src/application/controllers/AssetTypes.php
@@ -3,37 +3,43 @@ defined('BASEPATH') OR exit('No direct script access allowed');
 
 class AssetTypes extends CI_Controller {
 
-	private $page;
 	private $user_id;
 	private $user_role_id;
 
 	public function __construct() {
 		parent::__construct();
-		// check for user authorization
         $this->load->model('AssetTypes_model');
 		$this->load->helper("database");
 		$this->load->helper("general");
 		$this->load->helper("authorization");
-		$this->page = 'asset_groups';
 		$this->user_id = $this->session->userdata('id');
 		$this->user_role_id = $this->session->userdata('id');
-
+		$this->page = 'asset_groups';
+		if ( ! $this->session->userdata('id')) { // if the user is not logged in
+            redirect('unauthorized');
+		}
 		if ( ! is_authorized($this->user_role_id, $this->page)) {
-			echo 'unauthorized';
-			exit;
+			redirect('forbidden');
+		}
+
+		if ($this->uri->total_segments() > 1) {
+			if ( ! $this->uri->segment(2) == 'get_active') {
+				if ( ! has_write_access($this->user_role_id, $this->page)) {
+					redirect('forbidden');
+				}
+			}
 		}
 	}
 
-	public function index() {
-		log_message('debug', 'AssetTypes: index - in function');
-	}
+	// public function index() {
+	// 	log_message('debug', 'AssetTypes: index - in function');
+	// }
 
     public function add() {
 		log_message('debug', 'AssetTypes: add - in function');
 
 		if (!$this->input->is_ajax_request()) {
-            // echo $this->output_json(['unauthorized']);
-            exit;
+			redirect('forbidden');
         }
 		$this->form_validation->set_rules($this->AssetTypes_model->get_insert_rules());
 		if ($this->form_validation->run() == TRUE) {
@@ -63,8 +69,7 @@ class AssetTypes extends CI_Controller {
 		log_message('debug', 'AssetTypes: validate_add_name - in function');
 
 		if (!$this->input->is_ajax_request()) {
-			// echo $this->output_json(['unauthorized']);
-			exit;
+			redirect('forbidden');
 		}
 
 		$this->form_validation->set_rules(array($this->AssetTypes_model->get_insert_name_rules()));
@@ -83,8 +88,7 @@ class AssetTypes extends CI_Controller {
 		log_message('debug', 'AssetTypes: validate_add_rate - in function');
 
 		if (!$this->input->is_ajax_request()) {
-			// echo $this->output_json(['unauthorized']);
-			exit;
+			redirect('forbidden');
 		}
 
 		$this->form_validation->set_rules(array($this->AssetTypes_model->get_insert_rate_rules()));
@@ -101,9 +105,9 @@ class AssetTypes extends CI_Controller {
 
     public function edit() {
 		log_message('debug', 'AssetTypes: edit - in function');
+
 		if (!$this->input->is_ajax_request()) {
-            // echo $this->output_json(['unauthorized']);
-            exit;
+            redirect('forbidden');
         }
 
 		$this->form_validation->set_rules($this->AssetTypes_model->get_update_rules());
@@ -134,8 +138,7 @@ class AssetTypes extends CI_Controller {
 		log_message('debug', 'AssetTypes: validate_edit_name - in function');
 
 		if (!$this->input->is_ajax_request()) {
-            // echo $this->output_json(['unauthorized']);
-            exit;
+            redirect('forbidden');
         }
 
 		$this->form_validation->set_rules(array($this->AssetTypes_model->get_update_name_rules()));
@@ -152,11 +155,20 @@ class AssetTypes extends CI_Controller {
 
 	public function validate_edit_rate() {
 		log_message('debug', 'AssetTypes: validate_edit_rate - in function');
+
+		if (!$this->input->is_ajax_request()) {
+            redirect('forbidden');
+        }
+
 		$this->validate_add_rate();
 	}
 
     public function delete($id) {
 		log_message('debug', 'AssetTypes: delete - in function');
+
+		if (!$this->input->is_ajax_request()) {
+            redirect('forbidden');
+        }
 
         $this->AssetTypes_model->delete($id);
     }
@@ -164,25 +176,29 @@ class AssetTypes extends CI_Controller {
 	public function get_active() {
 		log_message('debug', 'AssetTypes: get_active - in function');
 
+		if (!$this->input->is_ajax_request()) {
+            redirect('forbidden');
+        }
+
 		$active_asset_types = $this->AssetTypes_model->get_active();
 		$json_asset_types = json_encode($active_asset_types);
 		echo $json_asset_types;
 	}
 
-	private function is_name_unique($name) {
+	public function is_name_unique($name) {
 		log_message('debug', 'AssetTypes: is_name_unique - in function');
 
 		return $this->AssetTypes_model->is_name_unique($name);
 	}
 
-	private function is_name_unique_not_different_from_current($name) {
+	public function is_name_unique_not_different_from_current($name) {
 		log_message('debug', 'AssetTypes: is_name_unique_not_different_from_current - in function');
 
 		$id = $this->input->post('id');
 		return $this->AssetTypes_model->is_name_unique_not_different_from_current($name, $id);
 	}
 
-	private function id_exists($id) {
+	public function id_exists($id) {
 		log_message('debug', 'AssetTypes: id_exists - in function');
 
 		return $this->AssetTypes_model->id_exists($id);
